pyproject.toml
[project]
name = "link-shortener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.31.0",
    "black>=25.11.0",
    "fastapi[standard]>=0.123.0",
    "pydantic-settings>=2.12.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "setuptools>=80.9.0",
    "sqlalchemy[asyncio]>=2.0.44",
]
